pyproject.toml
[project]
name = "super-bowl-ai-tracker"
version = "0.1.0"
description = "Super Bowl AI Tracker with Gemini-powered live commentary"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "jinja2>=3.1",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "pydantic>=2.6",
    "google-generativeai>=0.8.0",
]

[project.optional-dependencies]
test = ["pytest>=8.0"]

[tool.pytest.ini_options]
pythonpath = ["."]
